fix: improve timestamp precision in transcription and audio

+++ Client/SubtitlesApp.MAUI/Platforms/Android/FfmpegNative/FfmpegNativeService.cs
@@ -67,8 +67,8 @@ public partial class FfmpegNativeService : IAudioExtractor
         var ffmpegTask = Task.Run(() =>
         {
             var exitCode = FfmpegNativeWrapper.extract_audio(
-                (int)startTime.TotalSeconds,
-                (int)endTime.TotalSeconds,
+                startTime.TotalSeconds,
+                endTime.TotalSeconds,
                 sourcePath,
                 16000,
                 AudioFormats.Wave,
